// src/main/resources/fleet-core-frontend/src/app/components-2/auth/dto/Driver.ts
export class Driver{
    id: number;
    firstName: string;
    lastName: string;
    phoneNumber: string;
    email: string;
    role: string;
    imageData: string;
    organisationName: string;
    ratePerKilometer: number;
    licenseExpiryDate: Date;
    yearsOfExperience: number;
    totalKilometersDriven: number;
}