// src/main/resources/fleet-core-frontend/src/app/components-2/auth/dto/DecodedTokenDriver.ts
export class DecodedTokenDriver{
    id: number;
    firstName: string;
    lastName: string;
    phoneNumber: string;
    email: string;
    role: string;
    imageData: string;
    ratePerKilometer: number;
    licenseExpiryDate: Date;
    yearsOfExperience: number;
    totalKilometersDriven: number;
    organisationName: string;
}